move plugin detection to execution phase
